requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pandas
openai
supabase
python-dotenv
langchain
scikit-learn
flask
flask_cors
langchain_community
langchain-core
google-generativeai
langchain-google-genai
google-ai-generativelanguage==0.6.15

supabase
requests
werkzeug


openpyxl

psycopg2